+++ inflow/src/views/hr-management/vacation/VacationPaymentView.vue
@@ -1,68 +1,8 @@
 <template>
-  <ArticleItem
-    class="search-emp"
-    h="calc(100% - 23.6rem)"
-    w="32rem"
-    bgc="#f8f8f8"
-  >
-    <FlexItem
-      class="search-bar"
-      fld="row"
-      w="100%"
-      fs="1.5rem"
-      bgc="#fff"
-      b="0.6px solid #ccc"
-      br="0.3rem"
-    >
-      <input
-        name="search-emp-input"
-        type="text"
-        placeholder="사원코드 및 사원명으로 검색"
-      />
-      <SearchButton
-        h="4rem"
-        w="4rem"
-        bgc="#E6E6E6"
-        c="#5a5a5a"
-        fs="1.7rem"
-        br="0rem 0.3rem 0.3rem 0rem"
-      ></SearchButton>
-    </FlexItem>
-    <UlItem
-      v-if="!isEmpty"
-      class="emp-list"
-      fld="column"
-      h="calc(100% - 4rem)"
-      w="100%"
-    >
-      <LiItem
-        v-for="(item, index) in employeeList"
-        :key="index"
-        class="emp-item"
-        w="100%"
-        h="10rem"
-        bgc="#fff"
-        b="0.6px solid #ccc"
-        br="0.3rem"
-        @click="handleSelected(item)"
-      >
-        <span class="emphasize">{{ item.employee_name }}</span>
-        <span class="normal">{{ item.employee_number }}</span>
-        <span class="normal">{{ item.department_path }}</span>
-      </LiItem>
-    </UlItem>
-    <FlexItem
-      v-if="isEmpty"
-      class="empty-message"
-      fld="row"
-      h="6rem"
-      w="100%"
-      fs="1.4rem"
-    >
-      검색된 사원이 없습니다.
-    </FlexItem>
-  </ArticleItem>
-  <SectionItem class="payment-section" h="100%" w="calc(100% - 48rem)">
+  <SearchEmployeeComponent
+    @employee-selected="handleSelected"
+  ></SearchEmployeeComponent>
+  <SectionItem class="payment-section" h="100%" w="calc(100% - 52rem)">
     <CommonArticle label="휴가 지급">
       <FlexItem
         class="profile"
@@ -139,11 +79,7 @@
 </template>
 
 <script setup>
-import SearchButton from '@/components/buttons/SearchButton.vue';
-import ArticleItem from '@/components/semantic/ArticleItem.vue';
 import FlexItem from '@/components/semantic/FlexItem.vue';
-import LiItem from '@/components/semantic/LiItem.vue';
-import UlItem from '@/components/semantic/UlItem.vue';
 import SectionItem from '@/components/semantic/SectionItem.vue';
 import CommonArticle from '@/components/common/CommonArticle.vue';
 import FigureItem from '@/components/semantic/FigureItem.vue';
@@ -153,34 +89,20 @@ import TableCell from '@/components/semantic/TableCell.vue';
 import ButtonItem from '@/components/semantic/ButtonItem.vue';
 import DropdownItem from '@/components/dropdowns/DropdownItem.vue';
 import DateDropDown from '@/components/dropdowns/DateDropDown.vue';
+import SearchEmployeeComponent from '@/components/common/SearchEmployeeComponent.vue';
 import { ref, onMounted } from 'vue';
-import { getEmployeesByKeywordOrDepartmentCode } from '@/api/department';
 import {
   getIrregularVacationPoliciesByYear,
   createVacation,
 } from '@/api/vacation';
 
-const employeeList = ref([]);
 const vacationPolicyList = ref([]);
 const dropdownVacationPolicyList = ref([]);
 const isEmpty = ref(true);
-const keyword = ref('');
 
 const selectedEmployee = ref(null);
 const selectedVacationPolicy = ref(null);
 const selectedDate = ref(null);
-
-const fetchSearchedEmployeeData = async (keyword) => {
-  const response = await getEmployeesByKeywordOrDepartmentCode(keyword);
-
-  if (response.success) {
-    employeeList.value = response.content;
-    isEmpty.value = employeeList.value.isEmpty ? true : false;
-  } else {
-    employeeList.value = [];
-    isEmpty.value = true;
-  }
-};
 
 const fetchVacationPolicyData = async (year) => {
   const response = await getIrregularVacationPoliciesByYear(year);
@@ -268,41 +190,12 @@ const handleOnclick = async () => {
 };
 
 onMounted(() => {
-  keyword.value = '홍길동';
   const year = new Date().getFullYear();
-  fetchSearchedEmployeeData(keyword.value);
   fetchVacationPolicyData(year);
 });
 </script>
 
 <style scoped>
-.search-emp {
-  position: fixed;
-  padding: 0.8rem;
-  top: 23.6rem;
-  left: 14rem;
-  z-index: 2;
-}
-
-.search-bar input {
-  height: 100%;
-  width: calc(100% - 3.6rem);
-  padding-left: 1rem;
-  border-right: 0.6px solid #ccc;
-  overflow: hidden;
-}
-
-.emp-list {
-  margin-top: 1.5rem;
-  gap: 1rem;
-  overflow-y: auto;
-}
-
-.emp-item {
-  padding: 1.2rem;
-  gap: 0.3rem;
-}
-
 .emphasize {
   font-size: 2.2rem;
   font-weight: 500;
